Fix ordering in colony-factory-upgrade test
Whew, this was a subtle one folks. So the beforeEach here used a
waitAll, which fired off four transactions, and waited for them to
be mined. This is fine for the other two tests, but here the ordering
of `rootColony.moveColonyFactoryStorage` and `rootColony.registerColonyFactory`
matters - the storage has to be moved from the old colonyFactory, and
so if the new colonyFactory has already been registered, it won't
work and the test will fail.

Keep fighting the good fight against flickery tests, people!