Fix nightly builds that run on geth
So the (first?) reason that our nightly builds were failing was because
`geth` and `testrpc` treat transactions that don't specify a gas amount
differently. In `testrpc`, such transactions are given over 4 million gas
to play with (in fact, the whole block size). In `geth`, they're
given 90000 gas. This is probably going to change in `testrpc` soon,
https://github.com/ethereumjs/testrpc/issues/130 , so we should bear
that in mind going forward, but hopefully these changes will deal with it.

Why has this come about? Because we're no longer declaring default gas
amounts where we import our contracts, in either our tests or on the
clientside.

Why not just go back to declaring contract-wide gas limits? We removed it
in https://github.com/JoinColony/colonyDapp/commit/1826c4620daf872362bfc876696b5968aaade4e1 .
I'm proposing we don't - and set the gas where the default is inadequate.
This is better for the user (they're not signing messages where a single
unexpected throw could cost them a dollar), and keeps us more in touch
with how much gas we're using.

Going forward, I might see if I can add an option to `ether-pudding` that
automatically estimates gas, but our client will still want to do it itself
to tell the user the transaction they're signing. It's also pointless
if `estimateGas` is inaccurate, which seems to be a problem in `testrpc`
(see https://github.com/ethereumjs/testrpc/issues/58) and maybe
even `geth`(!?)  (see https://github.com/ethereum/go-ethereum/issues/1590)