Remove RegisterRootColonyResolver
We... don't want it? As it stood, admins could change the address
of the rootColonyResolver contract, and start getting the
fees (which they could set!) from the activity in the contract.
We should never want to change this - it is set on deployment, and
points to the rootColonyResolver. If we need to change the root
colony, we have the ability to do that through the rootColonyResolver -
we should never need to change the address of the resolver contract
that colonies point at.